requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
chromadb>=0.4.0
pydantic>=2.0.0
pytest>=7.4.0